ocr tool finalized/requirement.txt.txt
Flask==3.1.0
paddlepaddle==3.0.0
paddleocr==2.10.0
pdf2image==1.17.0
opencv-python==4.11.0.86
matplotlib==3.10.0
